leadmanager/requirements.txt
asgiref==3.2.10
astroid==2.4.2
authy==2.2.6
Babel==2.8.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
Django==3.0.5
django-cors-headers==3.4.0
django-extra-fields==2.0.5
django-filter==2.3.0
django-phone-verify==2.0.0
django-phonenumber-field==4.0.0
django-rest-auth==0.9.5
django-rest-knox==4.1.0
django-rest-registration==0.5.6
django-rest-swagger==2.2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
djongo==1.3.3
drf-base64==2.0
idna==2.10
importlib-metadata==1.7.0
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
mongoengine==0.15.3
nexmo==2.4.0
openapi-codec==1.3.2
phonenumbers==8.12.6
Pillow==7.2.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.5.3
pymongo==3.7.2
python-dotenv==0.14.0
pytz==2018.5
requests==2.24.0
simplejson==3.17.2
six==1.11.0
sqlparse==0.2.4
toml==0.10.1
twilio==6.44.0
typed-ast==1.4.1
typing==3.7.4.3
uritemplate==3.0.0
urllib3==1.23
wincertstore==0.2
wrapt==1.12.1
zipp==3.1.0
